+++ packages/multi-backend/src/ambient.d.ts
@@ -2,6 +2,13 @@ declare module "dnd-multi-backend" {
     import { Backend } from 'dnd-core';
     import { BackendFactory } from 'dnd-core';
     import { DragDropManager } from 'dnd-core';
+    export module PreviewManager {
+        export function register(preview: BackendWatcher): void;
+        export function unregister(preview: BackendWatcher): void;
+    }
+    export interface BackendWatcher {
+        backendChanged(backend: Backend): void;
+    }
     export interface Transition {
         event: string;
         check: (event: Event) => boolean;
@@ -9,11 +16,21 @@ declare module "dnd-multi-backend" {
     export type BackendTransition = {
         backend: BackendFactory;
         transition: Transition;
+        options?: any;
         preview?: boolean;
+        skipDispatchOnTransition?: boolean,
     };
-    const MultiBackend: (
-        transition: { backends: BackendTransition[] }
-    ) => BackendFactory;
+    export module PreviewManager {
+        export function register(preview: BackendWatcher): void;
+        export function unregister(preview: BackendWatcher): void;
+    }
+    export interface BackendWatcher {
+        backendChanged(backend: Backend): void;
+    }
+    export interface MultiBackendExt {
+        previewEnabled(): boolean;
+    }
+    const MultiBackend: BackendFactory;
     export default MultiBackend;
     export const createTransition: (
         event: string,