+++ packages/multi-backend/src/preview.component.ts
@@ -7,9 +7,16 @@ import {
     ChangeDetectionStrategy
 } from "@angular/core";
 import { SkyhookDndService, DRAG_DROP_MANAGER } from "@angular-skyhook/core";
-import { DragDropManager } from "dnd-core";
+import { DragDropManager, Backend } from "dnd-core";
 // @ts-ignore
 import { Observable } from 'rxjs';
+import { BehaviorSubject } from 'rxjs';
+import { PreviewManager, BackendWatcher } from 'dnd-multi-backend';
+
+
+interface MultiBackendExt {
+    previewEnabled(): boolean;
+}
 
 export interface PreviewTemplateContext {
     /** same as type */
@@ -37,19 +44,19 @@ export interface PreviewTemplateContext {
 @Component({
     selector: "skyhook-preview",
     template: `
-    <ng-container *ngIf="collect$|async as c">
-        <skyhook-preview-renderer *ngIf="c.previewEnabled">
-        <ng-container *ngIf="c.isDragging" >
-            <ng-container
-                *ngTemplateOutlet="content; context: { $implicit: c.itemType, type: c.itemType, item: c.item }">
-            </ng-container>
+    <ng-container *ngIf="previewEnabled$ | async">
+        <skyhook-preview-renderer *ngIf="collect$ | async as c">
+            <ng-container *ngIf="c.isDragging" >
+                <ng-container
+                    *ngTemplateOutlet="content; context: { $implicit: c.itemType, type: c.itemType, item: c.item }">
+                </ng-container>
             </ng-container>
         </skyhook-preview-renderer>
     </ng-container>
     `,
     changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class SkyhookPreviewComponent {
+export class SkyhookPreviewComponent implements BackendWatcher {
     /** Disables the check for whether the current MultiBackend wants the preview enabled */
     @Input() allBackends = false;
 
@@ -59,6 +66,9 @@ export class SkyhookPreviewComponent {
 
     /** @ignore */
     private layer = this.skyhook.dragLayer();
+
+    /** @ignore */
+    previewEnabled$ = new BehaviorSubject(false);
 
     // we don't need all the fast-moving props here, so this optimises change detection
     // on the projected template's inputs (i.e. the context).
@@ -70,7 +80,6 @@ export class SkyhookPreviewComponent {
         item: monitor.getItem(),
         itemType: monitor.getItemType(),
         isDragging: monitor.isDragging(),
-        previewEnabled: this.isPreviewEnabled()
     }));
 
     /** @ignore */
@@ -80,11 +89,31 @@ export class SkyhookPreviewComponent {
     constructor(
         private skyhook: SkyhookDndService,
         @Inject(DRAG_DROP_MANAGER) private manager: DragDropManager,
-    ) {}
+    ) {
+        if (this.manager == null) {
+            this.warn(
+                "no drag and drop manager defined, are you sure you imported SkyhookDndModule?"
+            );
+        } else {
+            PreviewManager.register(this);
+        }
+    }
+
+    /** @ignore */
+    ngOnInit() {
+        // Have to do this after allBackends receives its value.
+        this.backendChanged(this.manager.getBackend());
+    }
+
+    /** @ignore */
+    backendChanged(backend: Backend) {
+        this.previewEnabled$.next(this.isPreviewEnabled(backend));
+    }
 
     /** @ignore */
     ngOnDestroy() {
         this.layer.unsubscribe();
+        PreviewManager.unregister(this);
     }
 
     /** @ignore */
@@ -96,17 +125,10 @@ export class SkyhookPreviewComponent {
     }
 
     /** @ignore */
-    isPreviewEnabled() {
+    isPreviewEnabled(backend: Backend & Partial<MultiBackendExt>) {
         if (this.allBackends) {
             return true;
         }
-        if (this.manager == null) {
-            this.warn(
-                "no drag and drop manager defined, are you sure you imported SkyhookDndModule?"
-            );
-            return false;
-        }
-        const backend = this.manager.getBackend() as any;
         if (backend == null) {
             this.warn(
                 "no drag and drop backend defined, are you sure you imported SkyhookDndModule.forRoot(backend)?"
@@ -114,7 +136,7 @@ export class SkyhookPreviewComponent {
             return false;
         }
         // for when you are not using dnd-multi-backend
-        if (backend.previewEnabled == null) {
+        if (typeof backend.previewEnabled !== 'function') {
             return true;
         }
         return backend.previewEnabled();